# openapi/openapi.yaml
openapi: "3.0.0"
info:
  version: 0.1.5
  title: CRM API
  description: A main CRM Api
  contact:
    name: API Team
servers:
  - url: "localhost:8080"
    description: Dev server
...
paths:
...
  /reminder:
    get:
      description: Get reminder
      tags:
        - reminder
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - count
                  - type
                  - status
                properties:
                  count:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReminderDTO"

    post:
      description: Create reminder
      tags:
        - reminder
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/ReminderCreateRequest"
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/UUIDResponse"
  /reminder/{UUID}:
    put:
      description: Update reminder
      tags:
        - reminder
      parameters:
        - $ref: "#/components/parameters/uuid"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/ReminderPutRequest"
      responses:
        200:
          description: Ok

    delete:
      description: Delete reminder
      tags:
        - reminder
      parameters:
        - $ref: "#/components/parameters/uuid"
      responses:
        200:
          description: Ok
  /reminder/{UUID}/status:
    patch:
      description: Change reminder status
      tags:
        - reminder
      parameters:
        - $ref: "#/components/parameters/uuid"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/StatusRequest"
      responses:
        200:
          description: Ok
...
components:
  parameters:
    uuid:
      name: UUID
      in: path
      required: true
      schema:
        type: string
        format: uuid
        x-oapi-codegen-extra-tags:
          validate: "uuid"
...
  schemas:
...
    StatusRequest:
      type: object
      required:
        - status
        - comment
      properties:
        status:
          type: integer
          x-oapi-codegen-extra-tags:
            validate: "gte=0,lte=20"
        comment:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "trim,min=0,max=300"
    UUIDResponse:
      type: object
      required:
        - uuid
      properties:
        uuid:
          format: uuid
          type: string

    # BankAccountDTO 
...
    ReminderDTO:
      x-go-type: dto.ReminderDTO
      x-go-type-import:
        name: ReminderDTO
        path: github.com/krisch/crm-backend/dto
      type: object
      required:
        - uuid
        - description
        - date_to
        - date_from
        - task_uuid
        - created_at
        - updated_at
        - status
      properties:
        uuid:
          type: string
          format: uuid
        description:
          type: string
        date_to:
          type: string
          format: date-time
        date_from:
          type: string
          format: date-time
        task_uuid:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReminderCreateRequest:
      type: object
      required:
        - task_uuid
        - description
        - type
      properties:
        task_uuid:
          type: string
          format: uuid
          x-oapi-codegen-extra-tags:
            validate: "uuid"
        description:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "trim,name,min=0,max=2000"
        type:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "trim,name,min=0,max=50"
        date_to:
          type: string
          format: date-time
        date_from:
          type: string
          format: date-time
        user_uuid:
          type: string
          format: uuid
    ReminderPutRequest:
      type: object
      required:
        - description
        - comment
        - type
      properties:
        description:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "trim,name,min=0,max=2000"
        type:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "trim,name,min=0,max=50"
        comment:
          type: string
          x-oapi-codegen-extra-tags:
            validate: "trim,min=0,max=5000"
        date_to:
          type: string
          format: date-time
        date_from:
          type: string
          format: date-time
        user_uuid:
          type: string
          format: uuid
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []